ISSUE #1552: Simplify activity deletion
