+++ src/Application/Import/DeleteActivities/DeleteActivities.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace App\Application\Import\DeleteActivities;
+
+use App\Infrastructure\CQRS\Command\DomainCommand;
+use Symfony\Component\Console\Output\OutputInterface;
+
+final readonly class DeleteActivities extends DomainCommand
+{
+    public function __construct(
+        private OutputInterface $output,
+    ) {
+    }
+
+    public function getOutput(): OutputInterface
+    {
+        return $this->output;
+    }
+}
