--- src/Domain/Activity/BestEffort/DeleteActivityBestEfforts/DeleteActivityBestEffortsCommandHandler.php
@@ -1,24 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Domain\Activity\BestEffort\DeleteActivityBestEfforts;
-
-use App\Domain\Activity\BestEffort\ActivityBestEffortRepository;
-use App\Infrastructure\CQRS\Command\Command;
-use App\Infrastructure\CQRS\Command\CommandHandler;
-
-final readonly class DeleteActivityBestEffortsCommandHandler implements CommandHandler
-{
-    public function __construct(
-        private ActivityBestEffortRepository $activityBestEffortRepository,
-    ) {
-    }
-
-    public function handle(Command $command): void
-    {
-        assert($command instanceof DeleteActivityBestEfforts);
-
-        $this->activityBestEffortRepository->deleteForActivity($command->getActivityId());
-    }
-}
