--- src/Domain/Activity/Lap/DeleteActivityLaps/DeleteActivityLapsCommandHandler.php
@@ -1,24 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Domain\Activity\Lap\DeleteActivityLaps;
-
-use App\Domain\Activity\Lap\ActivityLapRepository;
-use App\Infrastructure\CQRS\Command\Command;
-use App\Infrastructure\CQRS\Command\CommandHandler;
-
-final readonly class DeleteActivityLapsCommandHandler implements CommandHandler
-{
-    public function __construct(
-        private ActivityLapRepository $activityLapRepository,
-    ) {
-    }
-
-    public function handle(Command $command): void
-    {
-        assert($command instanceof DeleteActivityLaps);
-
-        $this->activityLapRepository->deleteForActivity($command->getActivityId());
-    }
-}
