--- src/Domain/Activity/Split/DeleteActivitySplits/DeleteActivitySplitsCommandHandler.php
@@ -1,24 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Domain\Activity\Split\DeleteActivitySplits;
-
-use App\Domain\Activity\Split\ActivitySplitRepository;
-use App\Infrastructure\CQRS\Command\Command;
-use App\Infrastructure\CQRS\Command\CommandHandler;
-
-final readonly class DeleteActivitySplitsCommandHandler implements CommandHandler
-{
-    public function __construct(
-        private ActivitySplitRepository $activitySplitRepository,
-    ) {
-    }
-
-    public function handle(Command $command): void
-    {
-        assert($command instanceof DeleteActivitySplits);
-
-        $this->activitySplitRepository->deleteForActivity($command->getActivityId());
-    }
-}
