--- src/Domain/Segment/SegmentEffort/DeleteActivitySegmentEfforts/SegmentEffortsWereDeleted.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Domain\Segment\SegmentEffort\DeleteActivitySegmentEfforts;
-
-use App\Infrastructure\Eventing\DomainEvent;
-
-class SegmentEffortsWereDeleted extends DomainEvent
-{
-}
